Create custom hook preview
